requirements.txt
Django==3.2.15
celery==4.4.5
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-filter==2.3.0
mysqlclient==2.0.1
daphne==3.0.2
django-prometheus==2.1.0
prometheus-client==0.11.0
djangorestframework==3.11.1
WSGIserver==1.3
requests==2.24.0